+++ elements/map/src/main.js
@@ -158,6 +158,7 @@ export class EOxMap extends LitElement {
       sync: { attribute: "sync", type: String },
       zoomExtent: { attribute: false, type: Array },
       terrain: { type: Boolean },
+      useHighLOD: { attribute: "use-high-lod", type: Boolean },
     };
   }
 
@@ -238,12 +239,22 @@ export class EOxMap extends LitElement {
    * @type {ProjectionLike}
    */
   #_olProjection = "EPSG:3857";
-
   /**
    * Internal property to  store the terrain option for the globe projection.
    * @type {boolean}
    */
   #terrain = false;
+
+  /**
+   * Stores the last 2D projection to switch back from globe view.
+   * @type {ProjectionLike}
+   */
+  last2dProjection = "EPSG:3857";
+
+  /**
+   * Whether use high level of details or not.
+   */
+  #useHighLOD = false;
 
   constructor() {
     super();
@@ -508,6 +519,23 @@ export class EOxMap extends LitElement {
         this,
       );
     }
+  }
+
+  /**
+   * Sets if the (default) higher LOD level for the globe will not be used. This can be used to increase the level of detail for the globe, but may have performance implications.
+   *
+   * @param {boolean} useHighLOD - The useHighLOD.
+   */
+  set useHighLOD(useHighLOD) {
+    this.#useHighLOD = useHighLOD;
+  }
+
+  /**
+   * Gets the current useHighLOD state of the map. This indicates whether the globe will use a higher level of detail by not reducing the LOD levels.
+   * @returns {boolean}
+   */
+  get useHighLOD() {
+    return this.#useHighLOD;
   }
 
   /**